refactor: .ts파일 --> type.ts파일로 수정, 매핑함수 분리(helper.ts),  공통 커스텀 훅 생성(EscapeClose)
