+++ src/components/Modals/domains/Card/helper.ts
@@ -1,0 +1,16 @@
+import { CardModalUI, CardServerResponse } from "./ModalCard.type";
+
+// 데이터 매핑 함수
+export function mapCardToModalUI(
+  card: CardServerResponse
+): CardModalUI {
+  return {
+    title: card.title,
+    description: card.description,
+    tags: card.tags ?? null,
+    dueDateText: card.dueDate ?? null,
+    assigneeName: card.assignee?.nickname ?? null,
+    assigneeProfileUrl: card.assignee?.profileImageUrl ?? null,
+    imageUrl: card.imageUrl ?? null,
+  };
+}
