+++ src/components/Modals/domains/Column/ModalColumn.type.ts
