+++ src/hooks/useEscapeClose.ts
@@ -1,0 +1,14 @@
+import { useEffect } from 'react';
+
+export function useEscapeClose(onClose: () => void) {
+  useEffect(() => {
+    const handleKeyDown = (e: KeyboardEvent) => {
+      if (e.key === 'Escape') {
+        onClose();
+      }
+    };
+
+    window.addEventListener('keydown', handleKeyDown);
+    return () => window.removeEventListener('keydown', handleKeyDown);
+  }, [onClose]);
+}
